requirements.txt
click>=8.0.0
requests>=2.25.0
python-dotenv>=0.19.0
pynacl>=1.4.0
keyring>=23.0.0
black
pytest
pytest-cov
pylint

